upgrade Rebar class
