+++ spec/elements/reinforcement/rebar_hook_spec.rb
@@ -1,0 +1,7 @@
+require 'spec_helper'
+require 'db/base'
+require 'elements/reinforcement/rebar_hook'
+
+RSpec.describe Elements::Reinforcement::RebarHook do
+  xit 'does something'
+end
