--- spec/elements/steel_reinforcement/rebar_hook_spec.rb
@@ -1,7 +1,0 @@
-require 'spec_helper'
-require 'db/base'
-require 'elements/steel_reinforcement/rebar_hook'
-
-RSpec.describe Elements::SteelReinforcement::RebarHook do
-  xit 'does something'
-end
